fix: use fromAsset
This fixes the warning:

objectVersion is not defined for S3Code.fromBucket().
The lambda will not be updated automatically if the code in the bucket is
updated.

This is because CDK/Cloudformation does not track changes on the source
S3 Bucket.

It is recommended to either use S3Code.fromAsset() instead or set objectVersion.